Overhauled tabbing file structure
Split different concerns into different files. The resulting
configuration is not ideal, but better than before.
