+++ server/src/tabbing/audioData.ts
@@ -1,0 +1,7 @@
+/* Utility class for storing data loaded from an audio file. */
+export default class AudioData {
+    time: number[];
+    frequency: number[];
+    confidence: number[];
+    peakAmplitude: number[];
+}
