+++ server/src/tabbing/stringFret.ts
@@ -1,0 +1,5 @@
+/* Utility class for storing a string-fret pair. */
+export default class StringFret {
+    stringIdx: number; // 1-6 (low-high)
+    fret: number; // 0-12
+}
